Split AssetNodeStatusContent into a separate file
