+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/types/AssetNode.types.ts
@@ -15,6 +15,15 @@ export type AssetNodeLiveFragment = {
     runId: string;
   }>;
   assetObservations: Array<{__typename: 'ObservationEvent'; timestamp: string; runId: string}>;
+  assetChecks: Array<{
+    __typename: 'AssetCheck';
+    severity: Types.AssetCheckSeverity;
+    executionForLatestMaterialization: {
+      __typename: 'AssetCheckExecution';
+      id: string;
+      status: Types.AssetCheckExecutionResolvedStatus;
+    } | null;
+  }>;
   freshnessInfo: {__typename: 'AssetFreshnessInfo'; currentMinutesLate: number | null} | null;
   staleCauses: Array<{
     __typename: 'StaleCause';
