+++ js_modules/dagster-ui/packages/ui-core/src/graphql/schema.graphql
@@ -687,6 +687,7 @@ type AssetNode {
   type: DagsterType
   hasMaterializePermission: Boolean!
   hasAssetChecks: Boolean!
+  assetChecks: [AssetCheck!]!
 }
 
 type MaterializationUpstreamDataVersion {
@@ -834,6 +835,48 @@ enum StaleCauseCategory {
   CODE
   DATA
   DEPENDENCIES
+}
+
+type AssetCheck {
+  name: String!
+  assetKey: AssetKey!
+  description: String
+  severity: AssetCheckSeverity!
+  executions(limit: Int!, cursor: String): [AssetCheckExecution!]!
+  executionForLatestMaterialization: AssetCheckExecution
+}
+
+enum AssetCheckSeverity {
+  WARN
+  ERROR
+}
+
+type AssetCheckExecution {
+  id: String!
+  runId: String!
+  status: AssetCheckExecutionResolvedStatus!
+  evaluation: AssetCheckEvaluation
+  timestamp: Float!
+}
+
+enum AssetCheckExecutionResolvedStatus {
+  IN_PROGRESS
+  SUCCEEDED
+  FAILED
+  EXECUTION_FAILED
+  SKIPPED
+}
+
+type AssetCheckEvaluation {
+  timestamp: Float!
+  targetMaterialization: AssetCheckEvaluationTargetMaterializationData
+  metadataEntries: [MetadataEntry!]!
+}
+
+type AssetCheckEvaluationTargetMaterializationData {
+  storageId: Int!
+  runId: String!
+  timestamp: Float!
 }
 
 enum EvaluationErrorReason {
@@ -1157,6 +1200,7 @@ type AssetCheckEvaluationEvent implements MessageEvent & StepEvent {
   evaluation: AssetCheckEvaluation!
 }
 
+<<<<<<< HEAD
 type AssetCheckEvaluation {
   timestamp: Float!
   targetMaterialization: AssetCheckEvaluationTargetMaterializationData
@@ -1175,6 +1219,8 @@ enum AssetCheckSeverity {
   ERROR
 }
 
+=======
+>>>>>>> c18d8b9ed1 (Split AssetNodeStatusContent into a separate file)
 type PipelineRunLogsSubscriptionFailure {
   message: String!
   missingRunId: String
@@ -3186,6 +3232,7 @@ type AssetChecks {
   checks: [AssetCheck!]!
 }
 
+<<<<<<< HEAD
 type AssetCheck {
   name: String!
   assetKey: AssetKey!
@@ -3210,6 +3257,8 @@ enum AssetCheckExecutionResolvedStatus {
   SKIPPED
 }
 
+=======
+>>>>>>> c18d8b9ed1 (Split AssetNodeStatusContent into a separate file)
 type AssetCheckNeedsMigrationError implements Error {
   message: String!
 }
